Match 0x8004af18 and other related stuff (#222)
* Use user in docker container, install some packages

Create a user with the id, group id and name of the host. This way
the files produced by this container will be owned by the user.
Otherwise all the files produced by the conainer would be owner
by root which is annoying to work with.

Install gcc (needed by permuter) and progress (needed by objdiff)
Install pycparser (needed by permuter)

* match func_8004AF18

* cleanup func_8004AF18

* g_MapMsg_ColorTable

* func_8004AF18 => MapMsg_ParseAndRender

* cleanup

* GLYPH_TABLE_ASCII_OFFSET

* unblock Gfx_StringDraw

* D_80025D6C => g_MainFontWidthTable