Adding individual ovl tracking #1
Decomp.dev added a feature I requested of "sub-units" which allows to make to track individual parts of codes without cluttering the main unit list, thus allowing us to track individual overlays without cluttering main unit section with the 20+ overlays the game uses.

This will take a few commits as I have to implement the feature to the tracker and then figure out how to add the badges for each individual overlay.