bodyprog: func_80063A50 fix
In my preparation for Windows support I ran into a missmatch caused by this function, specifically `*(s32*)&ptr->field_164` was causing it, I asked for help and @emoose came up with two solutions: add .vx (as field_164 is a pointer to a SVECTOR) or use a macro they invented. I went for the second as they also mention that there are other instances that may benefit from it.