Repository Renovation: Added Ninja support, Makefile changes, Wiki created and Logo! (#458)
* Adding ninja support for the project attemp (commit 1)

* Adding ninja support for the project attemp (commit 2)

* Adding ninja support for the project attemp (commit 3)

I got it to work :3

* Adding ninja support to the project - Beautifying the code 1

* Adding ninja support to the project - Beautifying the code 2

* Adding ninja support to the project - Windows support 1

GOT IT TO COMPILE LET'S GOOOOOOOOOOOOOOOOO, now I'm going to test if this works in Linux/WSL and fix anything I broke, after that I will try to build bodyprog in Windows as it's not as simple like stream or main as in that case it requires iconv and I need to find a Windows counterpart for that

* Adding ninja support to the project - Windows support 2

EVERYTHING COMPILES NOW LET'S GOOOOOOOOOOOOOOOOOOOOOOOOOOO (still missing postbuild script so the FMVs overlay and engine matches ;-;)

* Adding ninja support to the project - Windows support 3

* Adding ninja support to the project - Fixes 1

* Adding ninja support to the project - Fixes 2

I forgot what I did. Now I'm investigating an issue that causes bodyprog to missmatch on windows and also make unable to insert missing rodata information in `bodyprog_800706E4.c`

* Adding ninja support to the project - Windows support 4

* Adding ninja support to the project - Windows support 4.1

da forgotten one

* Adding ninja support to the project - Windows support 5

Replaced postbuild batch script to Python script due performance

* Adding ninja support to the project - Fixes 3

Added an option individually work in each overlay or in a set overlays

* Adding ninja support to the project - Fixes 4

* Adding ninja support to the project - Fixes 5

The objdiff takeout

* Adding ninja support to the project - Fixes 5

Fixed objdiff ninja generation for Linux

* Adding ninja support to the project - Fixes 6

* Adding ninja support to the project - Features idk + 1

Added partial support for appending new overlays to the build files and regenerating them. Only supported for Objdiff compilation

* Adding ninja support to the project - Features idk + 2

The ninjas broke me update

* Adding ninja support to the project - Features idk + 2.1

* Adding ninja support to the project - Features idk + 3

* Adding ninja support to the project - Features idk + 4

* Adding ninja support to the project - Fixes 7

* Makefile adjustments - Commit 1

* Makefile adjustments - Commit 2

* Makefile adjustments - Commit 3

* Makefile adjustments - Commit 4

* Makefile adjustments - Commit 5

* Misc adjustments - Commit 1

* Misc adjustments - Commit 2

Added GCC compiler for m2ctx.py to work in windows. Hope we don't ended getting take down because of this stuff.

* Misc adjustments - Commit 3

* Misc adjustments - Commit 4

* Misc adjustments - Commit 5

* Misc adjustments - Commit 6

* Misc adjustments - Commit 7

Thanks emoose for pointing this out

* Misc adjustments - Commit 7.1