Bodyprog: Audio system recognition premature corrections
Found out that supposedly loads BGM key sounds in reality was loading the key notes (or the KDT file), also an irregularity with the first audio files made me take the decision of remove an enum. Also fixed an incorrect missmatch. This changes are premature as some comments require corrections and also fixed the incorrect match.